+++ .eslintrc.js
@@ -11,14 +11,12 @@ module.exports = {
     parser: '@babel/eslint-parser',
   },
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     'new-cap': ['error', { properties: false }],
     'tailwindcss/classnames-order': 'warn',
     'tailwindcss/no-custom-classname': 'warn',
     'tailwindcss/no-contradicting-classname': 'error',
     'vuejs-accessibility/label-has-for': 'off',
-    'global-require': 'off',
   },
   plugins: ['vue', 'tailwindcss'],
   settings: {
