+++ package.json
@@ -31,7 +31,6 @@
     "uuid": "^8.3.2",
     "vee-validate": "^4.5.8",
     "vue": "^3.0.0",
-    "vue-axios": "^3.4.1",
     "vue-cli-plugin-svg-sprite-loader": "^2.0.1",
     "vue-final-modal": "^3.4.3",
     "vue-router": "^4.0.0-0",
