+++ src/assets/scss/all.scss
@@ -1,0 +1,10 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@import url('https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css');
+@import 'swiper/scss';
+
+@import './help.scss';
+
+@import './componets.scss';
