+++ src/assets/scss/componets.scss
@@ -1,0 +1,5 @@
+@layer components {
+  .form-style {
+    @apply rounded border border-secondary-300 bg-secondary-50 py-2 px-3 text-secondary-500 focus:border-transparent focus:outline-none focus:ring-2 focus:ring-primary-700;
+  }
+}
