+++ src/components/modal/ModalCouponDetail.vue
@@ -34,15 +34,15 @@ export default {
     <div class="flex justify-between p-4">
       <div class="p-2 divide-y">
         <span class="block text-sm">優惠券名稱</span>
-        <h2 class="text-2xl font-medium">{{couponData?.title}}</h2>
+        <h2 class="text-2xl font-medium">{{ couponData?.title }}</h2>
       </div>
       <div class="p-2 divide-y">
         <span class="block text-sm">優惠券代碼</span>
-        <h2 class="text-2xl font-medium">{{couponData?.code}}</h2>
+        <h2 class="text-2xl font-medium">{{ couponData?.code }}</h2>
       </div>
       <div class="p-2 divide-y">
         <span class="block text-sm">優惠券折價</span>
-        <h2 class="text-2xl font-medium">{{couponData?.percent}}</h2>
+        <h2 class="text-2xl font-medium">{{ couponData?.percent }}</h2>
       </div>
     </div>
     <div class="flex justify-between p-4">
