+++ src/main.js
@@ -1,46 +1,65 @@
 import { createApp } from 'vue';
+
 import { createPinia } from 'pinia';
+
 import { vfmPlugin } from 'vue-final-modal';
-import axios from 'axios';
-import VueAxios from 'vue-axios';
+
 import VueSweetalert2 from 'vue-sweetalert2';
+
 import CKEditor from '@ckeditor/ckeditor5-vue';
+
 import { Roulette } from 'vue3-roulette';
+
 import AOS from 'aos';
 
-import App from '@/App.vue';
 import router from '@/router';
-import svgLoader from '@/utils/SvgLoader.vue';
+
+import SvgLoader from '@/utils/SvgLoader.vue';
+
 import AlertModal from '@/components/AlertModal.vue';
+
 import AppTitle from '@/utils/AppTitle.vue';
 
+import App from '@/App.vue';
+
 import 'aos/dist/aos.css';
+
 import 'sweetalert2/dist/sweetalert2.min.css';
+
 import 'bootstrap-icons/font/bootstrap-icons.css';
-import './style.scss';
+
+import '@/assets/scss/all.scss';
 
 const app = createApp(App);
 
 const pinia = createPinia();
 
-app.use(VueSweetalert2);
-app.provide('Swal', app.config.globalProperties.$swal);
+const requireAll = (requireContext) => requireContext.keys().forEach(requireContext);
 
-app.use(VueAxios, axios);
-app.provide('axios', app.config.globalProperties.axios);
-app.use(pinia);
-app.use(router);
-app.use(vfmPlugin);
-app.use(CKEditor);
+const req = require.context('@/assets/images/svg', true, /\.svg$/);
+
 app.AOS = new AOS.init({ disable: 'phone' });
 
-const requireAll = (requireContext) => requireContext.keys().forEach(requireContext);
-const req = require.context('@/assets/images/svg', true, /\.svg$/);
 requireAll(req);
 
-app.component('SvgLoader', svgLoader);
+app.provide('Swal', app.config.globalProperties.$swal);
+
+app.use(VueSweetalert2);
+
+app.use(pinia);
+
+app.use(router);
+
+app.use(vfmPlugin);
+
+app.use(CKEditor);
+
+app.component('SvgLoader', SvgLoader);
+
 app.component('AlertModal', AlertModal);
+
 app.component('AppTitle', AppTitle);
+
 app.component('AppRoulette', Roulette);
 
 app.mount('#app');
