+++ src/stores/order/index.js
@@ -1,8 +1,8 @@
-import { reactive, inject } from 'vue';
+import { reactive } from 'vue';
 import { defineStore } from 'pinia';
+import axios from 'axios';
 
 export const useOrderStore = defineStore('order', () => {
-  const axios = inject('axios');
   const baseUrl = process.env.VUE_APP_API_URL;
   const apiPath = process.env.VUE_APP_API_PATH;
   const orderResult = reactive({
