--- src/style.scss
@@ -1,69 +1,0 @@
-@import url('https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css');
-@import 'swiper/scss';
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-/* <transition name="fade">  */
-.fade-enter-active,
-.fade-leave-active {
-  transition: opacity 1s;
-}
-
-.fade-enter-from,
-.fade-leave-to {
-  opacity: 0;
-}
-
-.fade-enter-to,
-.fade-leave-from {
-  opacity: 1;
-}
-
-@layer components {
-  .form-style {
-    @apply rounded border border-secondary-300 bg-secondary-50 py-2 px-3 text-secondary-500 focus:border-transparent focus:outline-none focus:ring-2 focus:ring-primary-700;
-  }
-}
-
-.heroAnimate {
-  animation-name: heroAnimation;
-}
-@media screen and (orientation: landscape) {
-  @keyframes heroAnimation {
-    0% {
-      opacity: 0;
-      background-size: 150% auto;
-    }
-    25% {
-      opacity: 1;
-    }
-    75% {
-      opacity: 0;
-    }
-    100% {
-      opacity: 1;
-      background-size: 120% auto;
-    }
-  }
-}
-
-@media screen and (orientation: portrait) {
-  @keyframes heroAnimation {
-    0% {
-      opacity: 0;
-      background-size: auto 150%;
-    }
-    25% {
-      opacity: 1;
-    }
-    75% {
-      opacity: 0;
-    }
-    100% {
-      opacity: 1;
-      background-size: auto 120%;
-    }
-  }
-}
